Replace stdlib module with direct std imports
… except std::sync::Mutex, which there is a polyfill for.
