+++ tracing-core/src/callsite.rs
@@ -96,25 +96,24 @@
 //! [`Dispatch`]: crate::dispatcher::Dispatch
 //! [macros]: https://docs.rs/tracing/latest/tracing/#macros
 //! [instrument]: https://docs.rs/tracing/latest/tracing/attr.instrument.html
-use crate::stdlib::{
+
+use alloc::vec::Vec;
+use core::{
     any::TypeId,
     fmt,
     hash::{Hash, Hasher},
     ptr,
-    sync::{
-        atomic::{AtomicBool, AtomicPtr, AtomicU8, Ordering},
-        Mutex,
-    },
-    vec::Vec,
+    sync::atomic::{AtomicBool, AtomicPtr, AtomicU8, Ordering},
 };
+
+use self::dispatchers::Dispatchers;
 use crate::{
     dispatcher::Dispatch,
     lazy::Lazy,
     metadata::{LevelFilter, Metadata},
     subscriber::Interest,
+    sync::Mutex,
 };
-
-use self::dispatchers::Dispatchers;
 
 /// Trait implemented by callsites.
 ///
