+++ tracing-core/src/span.rs
@@ -1,7 +1,9 @@
 //! Spans represent periods of time in the execution of a program.
+
+use core::num::NonZeroU64;
+
 use crate::field::FieldSet;
 use crate::parent::Parent;
-use crate::stdlib::num::NonZeroU64;
 use crate::{field, Metadata};
 
 /// Identifies a span within the context of a subscriber.
